demos/use_cases/credit_risk_case_copilot/pyproject.toml
[project]
name = "credit-risk-case-copilot"
version = "0.1.0"
description = "Multi-agent Credit Risk Assessment System with Plano Orchestration"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn>=0.30.0",
    "pydantic>=2.11.7",
    "crewai>=0.80.0",
    "openai>=1.0.0",
    "httpx>=0.24.0",
    "streamlit>=1.40.0",
    "opentelemetry-api>=1.20.0",
    "opentelemetry-sdk>=1.20.0",
    "opentelemetry-exporter-otlp>=1.20.0",
    "opentelemetry-instrumentation-fastapi>=0.41b0",
    "python-dotenv>=1.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/credit_risk_demo"]
